Add explicit type conversions where necessary: media/
These are cases that are implicitly narrowing today, and must do so
explicitly in order to enable -Wc++11-narrowing.  No behavior change
intended.

Bug: 1216696
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2948073
Auto-Submit: Peter Kasting <pkasting@chromium.org>
Commit-Queue: Will Cassella <cassew@google.com>
Cr-Commit-Position: refs/heads/master@{#890792}

+++ media/cdm/cbcs_decryptor_unittest.cc
@@ -149,7 +149,8 @@ TEST_F(CbcsDecryptorTest, OneBlock) {
   DCHECK_EQ(kBlockSize, encrypted_block.size());
 
   // Only 1 subsample, all encrypted data.
-  std::vector<SubsampleEntry> subsamples = {{0, encrypted_block.size()}};
+  std::vector<SubsampleEntry> subsamples = {
+      {0, static_cast<uint32_t>(encrypted_block.size())}};
 
   auto encrypted_buffer = CreateEncryptedBuffer(
       encrypted_block, iv_, subsamples, EncryptionPattern(1, 9));
@@ -161,7 +162,8 @@ TEST_F(CbcsDecryptorTest, AdditionalData) {
   DCHECK_EQ(kBlockSize, encrypted_block.size());
 
   // Only 1 subsample, all encrypted data.
-  std::vector<SubsampleEntry> subsamples = {{0, encrypted_block.size()}};
+  std::vector<SubsampleEntry> subsamples = {
+      {0, static_cast<uint32_t>(encrypted_block.size())}};
 
   auto encrypted_buffer = CreateEncryptedBuffer(
       encrypted_block, iv_, subsamples, EncryptionPattern(1, 9));
@@ -191,7 +193,8 @@ TEST_F(CbcsDecryptorTest, DifferentPattern) {
   DCHECK_EQ(kBlockSize, encrypted_block.size());
 
   // Only 1 subsample, all encrypted data.
-  std::vector<SubsampleEntry> subsamples = {{0, encrypted_block.size()}};
+  std::vector<SubsampleEntry> subsamples = {
+      {0, static_cast<uint32_t>(encrypted_block.size())}};
 
   auto encrypted_buffer = CreateEncryptedBuffer(
       encrypted_block, iv_, subsamples, EncryptionPattern(1, 0));
@@ -203,7 +206,8 @@ TEST_F(CbcsDecryptorTest, EmptyPattern) {
   DCHECK_EQ(kBlockSize, encrypted_block.size());
 
   // Only 1 subsample, all encrypted data.
-  std::vector<SubsampleEntry> subsamples = {{0, encrypted_block.size()}};
+  std::vector<SubsampleEntry> subsamples = {
+      {0, static_cast<uint32_t>(encrypted_block.size())}};
 
   // Pattern 0:0 treats the buffer as all encrypted.
   auto encrypted_buffer = CreateEncryptedBuffer(
@@ -216,7 +220,8 @@ TEST_F(CbcsDecryptorTest, PatternTooLarge) {
   DCHECK_EQ(kBlockSize, encrypted_block.size());
 
   // Only 1 subsample, all encrypted data.
-  std::vector<SubsampleEntry> subsamples = {{0, encrypted_block.size()}};
+  std::vector<SubsampleEntry> subsamples = {
+      {0, static_cast<uint32_t>(encrypted_block.size())}};
 
   // Pattern 100:0 is too large, so decryption will fail.
   auto encrypted_buffer = CreateEncryptedBuffer(
@@ -239,7 +244,8 @@ TEST_F(CbcsDecryptorTest, BadSubsamples) {
   auto encrypted_block = Encrypt(one_block_, *key_, iv_);
 
   // Subsample size > data size.
-  std::vector<SubsampleEntry> subsamples = {{0, encrypted_block.size() + 1}};
+  std::vector<SubsampleEntry> subsamples = {
+      {0, static_cast<uint32_t>(encrypted_block.size() + 1)}};
 
   auto encrypted_buffer = CreateEncryptedBuffer(
       encrypted_block, iv_, subsamples, EncryptionPattern(1, 0));
@@ -249,7 +255,8 @@ TEST_F(CbcsDecryptorTest, BadSubsamples) {
 TEST_F(CbcsDecryptorTest, InvalidIv) {
   auto encrypted_block = Encrypt(one_block_, *key_, iv_);
 
-  std::vector<SubsampleEntry> subsamples = {{0, encrypted_block.size()}};
+  std::vector<SubsampleEntry> subsamples = {
+      {0, static_cast<uint32_t>(encrypted_block.size())}};
 
   // Use an invalid IV for decryption. Call should succeed, but return
   // something other than the original data.
@@ -262,7 +269,8 @@ TEST_F(CbcsDecryptorTest, InvalidIv) {
 TEST_F(CbcsDecryptorTest, InvalidKey) {
   auto encrypted_block = Encrypt(one_block_, *key_, iv_);
 
-  std::vector<SubsampleEntry> subsamples = {{0, encrypted_block.size()}};
+  std::vector<SubsampleEntry> subsamples = {
+      {0, static_cast<uint32_t>(encrypted_block.size())}};
 
   // Use a different key for decryption. Call should succeed, but return
   // something other than the original data.
@@ -276,7 +284,8 @@ TEST_F(CbcsDecryptorTest, InvalidKey) {
 TEST_F(CbcsDecryptorTest, PartialBlock) {
   // Only 1 subsample, all "encrypted" data. However, as it's not a full block,
   // it will be treated as unencrypted.
-  std::vector<SubsampleEntry> subsamples = {{0, partial_block_.size()}};
+  std::vector<SubsampleEntry> subsamples = {
+      {0, static_cast<uint32_t>(partial_block_.size())}};
 
   auto encrypted_buffer = CreateEncryptedBuffer(partial_block_, iv_, subsamples,
                                                 EncryptionPattern(1, 0));
@@ -291,7 +300,8 @@ TEST_F(CbcsDecryptorTest, SingleBlockWithExtraData) {
   auto expected_result = Combine({one_block_, partial_block_});
 
   // Only 1 subsample, all "encrypted" data.
-  std::vector<SubsampleEntry> subsamples = {{0, encrypted_block.size()}};
+  std::vector<SubsampleEntry> subsamples = {
+      {0, static_cast<uint32_t>(encrypted_block.size())}};
 
   auto encrypted_buffer = CreateEncryptedBuffer(
       encrypted_block, iv_, subsamples, EncryptionPattern(1, 0));
@@ -300,7 +310,8 @@ TEST_F(CbcsDecryptorTest, SingleBlockWithExtraData) {
 
 TEST_F(CbcsDecryptorTest, SkipBlock) {
   // Only 1 subsample, but all unencrypted data.
-  std::vector<SubsampleEntry> subsamples = {{one_block_.size(), 0}};
+  std::vector<SubsampleEntry> subsamples = {
+      {static_cast<uint32_t>(one_block_.size()), 0}};
 
   auto encrypted_buffer = CreateEncryptedBuffer(one_block_, iv_, subsamples,
                                                 EncryptionPattern(1, 0));
